Add retirement partner reporting queue and APIs
